[MOV] auth_totp{_mail{_enforce}}: merge module
This commit will merge the auth_totp_mail and auth_totp_enforce
modules into auth_totp.

Task-3349291
