+++ addons/auth_totp/models/__init__.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import auth_totp_rate_limit_log
 from . import auth_totp
 from . import ir_http
+from . import res_config_settings
 from . import res_users
 from . import totp
