+++ addons/auth_totp/tests/__init__.py
@@ -1,1 +1,2 @@
+from . import test_notify_security_update_totp
 from . import test_totp
