+++ addons/auth_totp/views/res_config_settings_views.xml
@@ -3,7 +3,7 @@
     <data>
 
         <record id="res_config_settings_view_form" model="ir.ui.view">
-            <field name="name">res.config.settings.view.form.inherit.auth_totp_mail_enforce</field>
+            <field name="name">res.config.settings.view.form.inherit.auth_totp</field>
             <field name="model">res.config.settings</field>
             <field name="priority" eval="40"/>
             <field name="inherit_id" ref="base_setup.res_config_settings_view_form"/>
